fix: supported protocols is only available since API v5
